==> 1. Introduction to Fast F1.ipynb <==
# INTRODUCTION TO FAST-F1 PACKAGE

import fastf1
from fastf1.core import Telemetry

import logging
# Get the FastF1 logger and set its level to ERROR
fastf1_logger = logging.getLogger('fastf1')
fastf1_logger.setLevel(logging.ERROR)


fastf1.get_session(year, gp, identifier=None, *, backend=None, force_ergast=False)
* session name abbreviation: 'FP1', 'FP2', 'FP3', 'Q', 'S', 'SS', 'R'

* full session name: 'Practice 1', 'Practice 2', 'Practice 3', 'Sprint', 'Sprint Shootout', 'Qualifying', 'Race'; provided names will be normalized, so that the name is case-insensitive

* number of the session: 1, 2, 3, 4, 5

session = fastf1.get_session(2023, 'Bahrain','Race')
session.event['EventName']

# Some Session Attributes

session.load()
session.results

driver_number = session.drivers
print(driver_number)

# Issued by race control officials during the event.
session.race_control_messages

session.weather_data

## CAR DATA (TELEMETRY)
	
Dictionary of car telemetry (Speed, RPM, etc.) as received from the api by car number (where car number is a string and the telemetry is an instance of Telemetry)

Car data Refers to telemetry data collected directly from the car's sensors and systems. Essential for analyzing the performance and behavior of the car itself.

* RPM: Revolutions Per Minute, the engine speed of the car's motor.
* Speed: Refers to the velocity of the car(km/h).
* nGear: Gear number the car is currently in.
* Throttle: The percentage of throttle pedal pressure applied by the driver, typically ranging from 0% (no throttle) to 100% (full throttle).
* Brake: Indicates whether the brakes are being applied (True) or not (False).
* DRS: Drag Reduction System, an aerodynamic device that reduces drag and increases speed on straights. It indicates whether the DRS is activated or not.
* Source: How the telemetry data was generated or obtained, such as directly from the car's sensors (car), positional data (pos), or through interpolation (interpolated).

session.car_data

## POS DATA (TELEMETRY)

Dictionary of car position data as received from the api by car number (where car number is a string and the telemetry is an instance of Telemetry)

Data (car's coordinates (X, Y, Z) on the track, status (on-track or off-track) etc) related to the position and status of the car on the track. Crucial for tracking the car's movement on the circuit, understanding its trajectory, adherence to the racing line, and monitoring any deviations from the track boundaries.

* Status: Status of the car's position, on track (OnTrack) or off track (OffTrack).

* X: The horizontal position of the car on the track, often measured in tenths of a meter.

* Y: The lateral position of the car on the track, typically measured in tenths of a meter.

* Z: The vertical position of the car relative to the track surface, usually measured in tenths of a meter.

* Source: Specifies how the positional data was obtained or generated, such as directly from the car's sensors or through interpolation.

session.pos_data

# LAP

laps = session.laps

## TELEMETRY DATA
Telemetry data for all laps in self

Telemetry data is the result of merging the returned data from get_car_data() and get_pos_data(). This means that telemetry data at least partially contains interpolated values! Telemetry data additionally already has computed channels added (e.g. Distance).

This method is provided for convenience and compatibility reasons. But using it does usually not produce the most accurate possible result. It is recommended to use get_car_data() or get_pos_data() when possible. 


**Telemetry can only be returned if self contains laps of one driver only.**

laps_driver_44 = laps.pick_driver('44')  # Hamilton
telemetry_driver_44 = laps_driver_44.get_telemetry()
print(telemetry_driver_44.columns)
telemetry_driver_44

## CAR DATA
Car data for all laps in self

Slices the car data in Session.car_data using this set of laps and returns the result.

The data returned by this method does not contain computed telemetry channels. The can be added by calling the appropriate add_*() method on the returned telemetry object..

car_driver_44 = laps_driver_44.get_car_data()
print(car_driver_44.columns)
car_driver_44

# POS DATA
Pos data for all laps in self

pos_driver_44 = laps_driver_44.get_pos_data()
print(pos_driver_44.columns)
pos_driver_44
