==> 3. Fetching Team Radio.ipynb <==
# 3. FETCHING TEAM RADIO (AUDIO) DATA

For now, we will be working on the 2023 season.

Track visualization practices were applied on Max Verstappen's performance in the 23-Bahrain Grand Prix. Let us understand the main source of team radio data and work on Verstappen's communication types, in this particular race.

### 3.1 REACHING THE JSON CONTENT

import fastf1
import os
import json
import requests
import urllib.request as request
import logging
from json import JSONDecodeError
import pandas as pd
import numpy as np
import speech_recognition as sr
from IPython.display import Audio 
import torch
import whisper
import tempfile
import soundfile as sf  
import random
import pickle
#import torchaudio

DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
parent_dir = os.path.dirname(os.path.realpath("3. Fetching Team Radio.ipynb")) 
data_path = os.path.join(parent_dir, 'Data Objects')

The 2023 Emilia Romagna GP was cancelled due to flood.

race_schedule_2023 = fastf1.get_event_schedule(2023)
race_schedule_2023

race_schedule_2022 = fastf1.get_event_schedule(2022)
race_schedule_2022

The link of the JSON file that keeps all audio data from a specific race:

livetiming.formula1.com/static/2023/2023-02-25_Bahrain_Grand_Prix/2023-02-25_Race/TeamRadio.jsonStream

def get_team_radio(location):
    logging.basicConfig(format='%(asctime)s - %(message)s', level=logging.INFO)
    url = location
    logging.info('Attempting to retrieve data from URL: %s', url)
    try:
        r = requests.get(location)
        r.encoding = 'utf-8-sig'
        out = json.loads(r.text)
        output = json.dumps(out, indent=4)
        logging.info('Data successfully retrieved from URL: %s', url)
        #return output
        return json.loads(output)
    except requests.RequestException as e:
        logging.error('Error retrieving data from URL: %s', url)
        logging.error(e)
        return None
    except JSONDecodeError:
        logging.warning('Failed to decode JSON data from URL: %s', url)
        return None

bahrain_location = "https://livetiming.formula1.com/static/2023/2023-03-05_Bahrain_Grand_Prix/2023-03-05_Race/TeamRadio.json"
data = get_team_radio(bahrain_location)
print("JSON data:", data) if data else print("Failed to retrieve JSON data.")


Let us automize the way of creating the link for a race other than Bahrain.

The formula goes as follows:

"https://livetiming.formula1.com/static/ **year** /**date**_**eventName**/**date**_Race/TeamRadio.json"

team_radio_links = {}
for i in race_schedule_2023[1:].index:
    date = race_schedule_2023.EventDate[i] #format %Y-%m-%d %H:%M:%S.%f
    event_name = race_schedule_2023.EventName[i].replace(" ", "_")
    country = race_schedule_2023.Country[i]
    
    team_radio_links[event_name] = "https://livetiming.formula1.com/static/" + date.strftime('%Y') + "/" + date.strftime('%Y-%m-%d') + "_" + event_name + "/" + date.strftime('%Y-%m-%d') + "_Race/TeamRadio.json"
    
print(team_radio_links)

Apply the same procedure for the 2022 season.

team_radio_links22 = {}
for i in race_schedule_2022[2:].index: #First two are pre-season testings
    date = race_schedule_2022.EventDate[i] #format %Y-%m-%d %H:%M:%S.%f
    event_name = race_schedule_2022.EventName[i].replace(" ", "_")
    country = race_schedule_2022.Country[i]
    
    team_radio_links22[event_name] = "https://livetiming.formula1.com/static/" + date.strftime('%Y') + "/" + date.strftime('%Y-%m-%d') + "_" + event_name + "/" + date.strftime('%Y-%m-%d') + "_Race/TeamRadio.json"
    
print(team_radio_links22)

Now, all of the team radio JSON sources are in the **team_radio_links** dictionary.

Below, we run a small test to see all our links are correct and reachable. We can see that all the content is reached, successfully.

Now it is time to organize the data

for i in team_radio_links.keys():
    print(i)
    data = get_team_radio(team_radio_links[i])
    print("") if data else print("ERROR! INDEX:",i)

for i in team_radio_links22.keys():
    print(i)
    data = get_team_radio(team_radio_links22[i])
    print("") if data else print("ERROR! INDEX:",i)

Grouping the team radio so that all conversations from one driver is together.

def group_by_driver(data): 
    grouped_data = {}
    for item in data:
        racing_number = item['RacingNumber']
        if racing_number not in grouped_data:
            grouped_data[racing_number] = []
        grouped_data[racing_number].append(item)
        
    grouped_data = dict(sorted(grouped_data.items()))
    return grouped_data

def dict2df(dict):
    dfs = []
    for racing_number, data_list in dict.items():
        df = pd.DataFrame(data_list)
        dfs.append(df)
    
    result = pd.concat(dfs, keys=dict.keys())
    return result

content = get_team_radio(team_radio_links['Bahrain_Grand_Prix'])
data = content['Captures']
grouped_data = group_by_driver(data)

dict2df(grouped_data)

dict2df(grouped_data).loc['1'] # Driver with racing number '1' (Max Verstappen)

Now we have a basic understanding of how our team radio data should be storaged and organized, we need to apply the same procedure for all countries.

We also convert the audio links into the appropriate form.


**link we need:** https://livetiming.formula1.com/static/2023/2023-03-05_Bahrain_Grand_Prix/2023-03-05_Race/TeamRadio/GUAZHO01_24_20230305_163627.mp3

**split**
https://livetiming.formula1.com/static/2023/2023-03-05_Bahrain_Grand_Prix/2023-03-05_Race/TeamRadio.json
->
https://livetiming.formula1.com/static/2023/2023-03-05_Bahrain_Grand_Prix/2023-03-05_Race/ + TeamRadio/GUAZHO01_24_20230305_163627.mp3

team_radio = {}
for k,v in team_radio_links.items():
    content = get_team_radio(team_radio_links[k])
    data = content['Captures']
    
    for capture in data:
        # CREATE NEW LINK
        path = capture['Path']
        path_new = v.split("TeamRadio.json",1)[0] + path #creating a new link        
        capture['Path New'] = path_new
        
        # CREATE TIMESTAMP
        ts = pd.Timestamp(capture['Utc']).round('1ms')
        ts_new = pd.Timestamp(ts.strftime("%Y-%m-%d %H:%M:%S.%f"))
        capture['Date'] = ts_new
    
    grouped_data = group_by_driver(data)
    team_radio[k] = dict2df(grouped_data)

team_radio22 = {}
for k,v in team_radio_links22.items():
    content = get_team_radio(team_radio_links22[k])
    data = content['Captures']
    
    for capture in data:
        # CREATE NEW LINK
        path = capture['Path']
        path_new = v.split("TeamRadio.json",1)[0] + path #creating a new link        
        capture['Path New'] = path_new
        
        # CREATE TIMESTAMP
        ts = pd.Timestamp(capture['Utc']).round('1ms')
        ts_new = pd.Timestamp(ts.strftime("%Y-%m-%d %H:%M:%S.%f"))
        capture['Date'] = ts_new
    
    grouped_data = group_by_driver(data)
    team_radio22[k] = dict2df(grouped_data)

print("COUNTRIES IN THE DICTIONARY\n",team_radio.keys())
print("\n\nCOLUMNS\n",team_radio['Saudi_Arabian_Grand_Prix'].columns)

team_radio['Bahrain_Grand_Prix']

team_radio22['Bahrain_Grand_Prix']

team_radio['Mexico_City_Grand_Prix']

Now it is time to import the mp3 audio data.

### 3.2 IMPORTING AUDIO DATA

audio_url = team_radio['Bahrain']['Path New'].iloc[3]
print(audio_url) #Max Verstappen's 8 seconds long radio.

Audio(audio_url, autoplay=True)

def download_audio(url):
    response = requests.get(url)
    if response.status_code == 200:
        return response.content  # Return the audio data
    else:
        print("Failed to download audio")
        return None

Python's Speech Recognition module.
> https://pypi.org/project/SpeechRecognition/

Speech recognition engine/API support:
- CMU Sphinx (works offline)
- Google Speech Recognition
- Google Cloud Speech API
- Wit.ai
- Microsoft Azure Speech
- Microsoft Bing Voice Recognition (Deprecated)
- Houndify API
- IBM Speech to Text
- Snowboy Hotword Detection (works offline)
- Tensorflow
- Vosk API (works offline)
- OpenAI whisper (works offline)
- Whisper API

### 3.2.1 Google Speech Recognition Model

def speech_to_text(audio_data, sample_rate, sample_width):
    recognizer = sr.Recognizer()
    # Converting the audio data to audio source
    audio_source = sr.AudioData(audio_data, sample_rate, sample_width)  #Sample_rate and sample_width

    # Speech to text conversion
    try:
        text = recognizer.recognize_google(audio_source)  #we are using Google Speech Recognition API
        return text
    except sr.UnknownValueError:
        print("Speech recognition could not understand audio")
        return None
    except sr.RequestError as e:
        print(f"Could not request results from Speech Recognition service; {e}")
        return None

audio_data = download_audio(audio_url)
if audio_data:
    result = speech_to_text(audio_data, 48000, 4)
    if result:
        print("Text from audio:", result)

Ideas to solve this issue:

- Verify sample rate and width (from the original audio data source)
- To void background noises. (this might be the main issue)
    - Noise reduction 
    - Filters: Low-pass filters, high-pass filters, band-stop filters
    - Preprocessing :??
- Additional parameters: Sensitivity threshold, noise suppression settings, language?

### 3.2.2 OpenAI's Whisper Model

def download_audio_new(url):
    response = requests.get(url)
    if response.status_code == 200:
        with tempfile.NamedTemporaryFile(delete=False) as tmp_file:
            tmp_file.write(response.content)
            tmp_file_name = tmp_file.name
        audio_data, _ = sf.read(tmp_file_name, dtype='float32')  # Read audio file as waveform
        return audio_data
    else:
        print("Failed to download audio")
        return None

def speech_to_text_whisper(url,model):
    audio_data = download_audio_new(url)
    results = model.transcribe(audio_data)
    return results['text']

model = whisper.load_model("base.en")
print(
    f"Model is {'multilingual' if model.is_multilingual else 'English-only'} "
    f"and has {sum(np.prod(p.shape) for p in model.parameters()):,} parameters."
)
options = whisper.DecodingOptions(language="en", without_timestamps=True)
options

print(speech_to_text_whisper(audio_url,model))

for country in team_radio.keys():
    print(f"{country:<20} : {len(team_radio[country]):<20}")

country = random.sample(list(team_radio.keys()),k=1)[0]
index = random.randint(0, 55)

audio_url = team_radio[country]['Path New'].iloc[index]

print(speech_to_text_whisper(audio_url,model))
Audio(audio_url,autoplay=True)

Open AI's Whisper model also seems to be not working well.

### 3.4 SAVE DICTIONARY

We must save the team radio-timestamps dictionary to our file so that we can import and use in the next phases of our project.

with open(os.path.join(data_path,'team_radio23.pkl'), 'wb') as f:
    pickle.dump(team_radio, f)

with open(os.path.join(data_path,'team_radio22.pkl'), 'wb') as f:
    pickle.dump(team_radio22, f)
