==> 4. Telemetry & Timestamps.ipynb <==
# 4. TELEMETRY & TIMESTAMPS

We will integrate our progress so far by visualizing the timestamp points on the track where team radio conversations take place for each driver, alongside analyzing the car's telemetry data at these specific locations.

### 4.1 CONSTRUCT DICTIONARY

import fastf1
import os
import pickle
import numpy as np
import pandas as pd

parent_dir = os.path.dirname(os.path.realpath("4. Telemetry & Timestamps.ipynb")) 
data_path = os.path.join(parent_dir, 'Data Objects')

with open(os.path.join(data_path, 'team_radio23.pkl'), 'rb') as f:
    team_radio = pickle.load(f)
    
print(len(team_radio.keys()),"races in 2023!")

with open(os.path.join(data_path, 'team_radio22.pkl'), 'rb') as f:
    team_radio22 = pickle.load(f)
    
print(len(team_radio22.keys()),"races in 2022!")

team_radio.keys()

team_radio22.keys()

Common races: 'Bahrain_Grand_Prix', 'Saudi_Arabian_Grand_Prix', 'Australian_Grand_Prix', 'Azerbaijan_Grand_Prix', 'Miami_Grand_Prix', 'Monaco_Grand_Prix', 'Spanish_Grand_Prix', 'Canadian_Grand_Prix', 'Austrian_Grand_Prix', 'British_Grand_Prix', 'Hungarian_Grand_Prix', 'Belgian_Grand_Prix', 'Dutch_Grand_Prix', 'Italian_Grand_Prix', 'Singapore_Grand_Prix', 'Japanese_Grand_Prix', 'United_States_Grand_Prix', 'Mexico_City_Grand_Prix', 'São_Paulo_Grand_Prix', 'Abu_Dhabi_Grand_Prix'

Before we continue, we need to be cautious of potential issues/missing from our data.

There might be certain races in which a driver is in the list of session drivers BUT not in the list of team radio drivers or vice-versa.

We need to detect these certain cases before we join 2 data objects (team radio timestamps & telemetry data) by driver number!

def analyse_season(dictt, year):
    i=1
    for k,v in dictt.items():
        session = fastf1.get_session(year, k,'Race') 
        session.load()
        
        drivers = sorted(session.drivers)
        rd = v.index.get_level_values(0).unique()
        
        different_drivers = [x for x in drivers if x not in set(rd)]
        different_drivers.append([x for x in rd if x not in set(drivers)])
        
        print(i," ",k,"\n")
        print("RACE \n",len(drivers)," ",drivers)
        print("RADIO \n",len(rd)," ",rd)
        if different_drivers !=[[]]: # if there is an inbalance between team radio drivers & session drivers in a race!
            print("\n Difference: ",different_drivers)
        
        i+=1

def driver_from_session(grand_prix,number):
    s= fastf1.get_session(2023, grand_prix,'Race') 
    s.load()
    print(s.get_driver(number))
    
def driver_from_radio(grand_prix,number):
    df = team_radio[grand_prix] 
    print(df.loc[(number, slice(None))]['Path New'][0])

analyse_season(team_radio22, 2022)

## 2022 SEASON
### SAUDI GP
**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '27', '3', '31', '4', '44', '47', '55', '6', '63', '77']
- 
**RADIO** 
- 19   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '27', '3',
       '31', '4', '44', '55', '6', '63', '77']

**Difference:** ['47']
- No radio for 	Mick Schumacher(47), but he raced.

### Miami GP
**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '3', '31', '4', '44', '47', '5', '55', '6', '63', '77']
- 
**RADIO**
- 18   ['1', '10', '14', '16', '18', '20', '22', '23', '3', '31', '4', '44',
       '47', '5', '55', '6', '63', '77']
       
### British GP 
**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '3', '31', '4', '44', '47', '5', '55', '6', '63', '77']
- 
**RADIO**
- 18   ['1', '10', '11', '14', '16', '18', '20', '22', '3', '31', '4', '44',
       '47', '5', '55', '6', '63', '77']

**Difference:** ['23', '24']

### Singapore GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '3', '31', '4', '44', '47', '5', '55', '6', '63', '77']

**RADIO**
 19   ['1', '10', '11', '16', '18', '20', '22', '23', '24', '3', '31', '4',
       '44', '47', '5', '55', '6', '63', '77']

**Difference:**  ['14']

### São Paulo GP 

**RACE** 
- 20   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '3', '31', '4', '44', '47', '5', '55', '6', '63', '77']
**RADIO** 
- 19   ['1', '10', '11', '14', '16', '18', '20', '22', '23', '24', '31', '4',
       '44', '47', '5', '55', '6', '63', '77']

**Difference:**  ['3']

### GPs WHERE ALL DRIVERS MATCH!

- Bahrain GP
- Australian GP
- Emilia Romagna GP
- Spanish GP
- Monaco GP
- Azerbaijan GP
- Canadian GP
- Austrian GP
- French GP
- Hungarian GP
- Belgian GP
- Italian GP
- Dutch GP
- Japanese GP
- United States GP
- Mexico City GP
- Abu Dhabi GP

analyse_season(team_radio, 2023)

## 2023 SEASON
### Canadian GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '21', '22', '23', '24', '27', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '31', '4', '44', '55', '63', '77', '81']

**Difference:** ['21']
- No radio for Nyck De Vries(21), but he raced.

### Hungarian GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 18   ['1', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '31', '4', '44', '55', '63', '77', '81']

**Difference:** ['10', '3']
- No radio for Pierre Gasly(10) & Daniel Ricciardo(3), but they raced.

### Dutch GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '31', '4', '40', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '27', '31', '4', '40', '44', '55', '63', '77', '81']

**Difference:** ['24']
- No radio for Guanyu Zhou(24), but he raced.

### Singapore GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '31', '4', '40', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '2', '20', '22', '23', '24', '27', '31', '4', '40', '44', '55', '63', '77', '81']

**Difference:** ['18']
- No radio for Lance Stroll(18), he did not race due to an injury.

### Qatar GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '31', '4', '40', '44', '63', '77', '81']

**Difference:** ['3', '55', '40']
- No radio for Lewis Hamilton(55), he DNF'ed.
- No radio for Daniel Ricciardo(3), he did not race due to an injury.

- There is radio available for Liam Lawson(40) but he is not on the driver list, he was a replacement of Ricciardo(3).
- Even though, Daniel Ricciardo(3) did not race, he is in the list of drivers.

### United States GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '21', '22', '23', '24', '27', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**Difference:** ['21', '3']
- Nyck De Vries(21) is in the driver list, even though he did not race.
- There is radio available for Daniel Ricciardo(3) but he is not on the driver list.
- Daniel Ricciardo(3) was a replacement for Nyck De Vries(21).

### São Paulo GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**Difference:** ['23']
- No radio for Alexander Albon(23), he DNF'ed.

### Las Vegas GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '44', '55', '63', '77', '81']

**Difference:** ['4']
- No radio for Lando Norris(4), he DNF'ed.

### Abu Dhabi GP

**RACE**
- 20   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '24', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**RADIO**
- 19   ['1', '10', '11', '14', '16', '18', '2', '20', '22', '23', '27', '3', '31', '4', '44', '55', '63', '77', '81']

**Difference:** ['24']
- No radio for Zhou Guanyu(24), but he raced.

### GPs WHERE ALL DRIVERS MATCH!

- Bahrain GP
- Saudi GP
- Australian GP
- Azerbaijan GP
- Miami GP
- Monaco GP
- Spanish GP
- Austrian GP
- British GP
- Belgian GP
- Italian GP
- Japanese GP
- Mexico City GP


def match_timestamps(date, telemetry):
    
    """
    Right join function where it returns the telemetry data of timestamps
    where radio conversation occurs.
    :param date: Team radio timestamps.
    :param telemetry: All timestamps available in telemetry data
    :return: data frame containing telemetry
    """
    
    dfs = []
    team_radio_ts = date
    telemetry_ts = telemetry['Date'].values 
    
    for ts in team_radio_ts:
        index = np.where(telemetry_ts == ts)[0]
        if len(index) > 0:
            df = telemetry.iloc[[index[0]], :]
            dfs.append(df)
        else:
            index = np.where(telemetry['Date'] < ts)[0]
            if len(index) > 0:
                df = telemetry.iloc[[index[-1]], :]
                dfs.append(df)
            else: 
                index = np.where(telemetry['Date'] > ts)[0]
                df = telemetry.iloc[[index[0]], :]
                dfs.append(df)

    result = pd.concat(dfs, ignore_index=True) 
    return result

def match_timestamps_reverse(date, telemetry):
    
    """
    :param date: Team radio timestamps.
    :param telemetry: All timestamps available in telemetry data
    :return: Index list of telemetry data where radio conversations occur.
    """

    index_list = []
    team_radio_ts = date
    telemetry_ts = telemetry['Date'].values 
    
    for ts in team_radio_ts:
        index = np.where(telemetry_ts == ts)[0]
        if len(index) > 0:
            index_list.append(index[0])
        else:
            index = np.where(telemetry['Date'] < ts)[0]
            if len(index) > 0:
                index_list.append(index[-1])
            else: 
                index = np.where(telemetry['Date'] > ts)[0]
                index_list.append(index[0])

    return index_list

def dict2df(dictionary):
    dfs = []
    for racing_number, data_list in dictionary.items():
        df = pd.DataFrame(data_list)
        dfs.append(df)
    
    result = pd.concat(dfs, keys=dictionary.keys())
    return result

correct_gps22= grand_prix = ["Bahrain_Grand_Prix","Australian_Grand_Prix","Emilia_Romagna_Grand_Prix","Spanish_Grand_Prix","Monaco_Grand_Prix","Azerbaijan_Grand_Prix","Canadian_Grand_Prix","Austrian_Grand_Prix","French_Grand_Prix","Hungarian_Grand_Prix","Belgian_Grand_Prix","Italian_Grand_Prix","Dutch_Grand_Prix","Japanese_Grand_Prix","United_States_Grand_Prix","Mexico_City_Grand_Prix","Abu_Dhabi_Grand_Prix"]

correct_gps23 = ['Bahrain_Grand_Prix','Saudi_Arabian_Grand_Prix','Australian_Grand_Prix','Azerbaijan_Grand_Prix','Miami_Grand_Prix','Monaco_Grand_Prix','Spanish_Grand_Prix','Austrian_Grand_Prix','British_Grand_Prix',
'Belgian_Grand_Prix','Italian_Grand_Prix','Japanese_Grand_Prix','Mexico_City_Grand_Prix']

def create_dict(team_radio_dict,gp, year):
    f1_dict = {}
    for k,v in team_radio_dict.items():
        # k is the country
        # v is the data frame of the related country with columns: 'Utc', 'RacingNumber', 'Path', 'Path New', 'Date'
        if np.isin(gp,k).any():
            session = fastf1.get_session(year, k,'Race') 
            session.load()
            
            drivers = sorted(session.drivers)
            
            #f1[k] = {}
            f1_dict[k] = {}
            for driver in drivers:
                df = v.loc[(driver, slice(None))]
                dates = df['Date'] #timestamps where team radio conversation occurs
                
                tel = session.laps.pick_driver(driver).get_telemetry()
                tel = tel[['Date','X','Y','Z','RPM','Brake','DRS','Throttle','Speed', 'nGear']] 
                
                index_list = match_timestamps_reverse(dates, tel)
                telemetry = tel.copy()
                telemetry.reset_index()
                telemetry.insert(len(telemetry.columns),"Team Radio",0)
                
                for i in index_list:
                    telemetry.loc[i,'Team Radio']=1
                    
                telemetry = telemetry.reset_index(drop=True)
                f1_dict[k][driver] = telemetry
                
    return f1_dict

v=team_radio['Dutch_Grand_Prix']

session = fastf1.get_session(2023, 'Dutch_Grand_Prix','Race') 
session.load()
df = v.loc[('1', slice(None))]
dates = df['Date'] #timestamps where team radio conversation occurs

tel = session.laps.pick_driver('1').get_telemetry()
tel = tel[['Date','X','Y','Z','RPM','Brake','DRS','Throttle','Speed', 'nGear']] 

tel

dates

index_list = match_timestamps_reverse(dates, tel)
telemetry = tel.copy()
telemetry.reset_index()
telemetry.insert(len(telemetry.columns),"Team Radio",0)

for i in index_list:
    telemetry.loc[i,'Team Radio']=1
    
telemetry = telemetry.reset_index(drop=True)
telemetry

telemetry.index[telemetry['Team Radio'] == 1].tolist()


tel.iloc[1315:,:]

telemetry.iloc[1315:,:]

f1_dict22 = create_dict(team_radio22,correct_gps22, 2022)

f1_dict23 = create_dict(team_radio,correct_gps23, 2023)

check and delete null rows!

def change_index(full_dict):
    """
    Changing the index of the dataframes (in the dictionary) to X-Y coordinates.
    """
    for gp,sub_dict in full_dict.items():
        for driver, df in sub_dict.items():
            df['X-Y'] = list(zip(df['X'], df['Y']))
            df.set_index('X-Y',inplace=True)

def organize_dict(full_dict):
    """
    Organizing the values of the dictionary; by changing the indexes, converting dictionary typed values to data frames
    and dropping null rows.
    """
    change_index(full_dict)
    new_dict = {}
    for gp, sub_dict in full_dict.items():
        df = dict2df(sub_dict)
        df = df.dropna()
        new_dict[gp] = df
    return new_dict

f1_22 = organize_dict(f1_dict22)
f1_23 = organize_dict(f1_dict23)

f1_22['Japanese_Grand_Prix']

f1_23['Japanese_Grand_Prix']

with open(os.path.join(data_path,'f1_22.pkl'), 'wb') as f:
    pickle.dump(f1_22, f)

with open(os.path.join(data_path,'f1_23.pkl'), 'wb') as f:
    pickle.dump(f1_23, f)
